+++ src/cpp/src/nn/layers/gnn/graph_sage_layer.cpp
@@ -72,7 +72,7 @@ torch::Tensor GraphSageLayer::forward(torch::Tensor inputs, DENSEGraph dense_gra
         }
     }
 
-    int64_t layer_offset = dense_graph.getLayerOffset();
+    int32_t layer_offset = dense_graph.getLayerOffset();
     torch::Tensor self_embs = inputs.narrow(0, layer_offset, inputs.size(0) - layer_offset);
 
     torch::Tensor outputs;
