+++ test/cpp/unit/testing_util.cpp
@@ -6,7 +6,7 @@
 int createTmpFile(std::string &filename) { return open(filename.c_str(), O_RDWR | O_CREAT, 0777); }
 
 torch::Tensor getRandTensor(int dim0_size, int dim1_size, torch::Dtype dtype, int max_val) {
-    if (dtype == torch::kInt32 || dtype == torch::kInt64) {
+    if (dtype == torch::kInt32 || dtype == torch::kInt32) {
         return torch::randint(max_val, {dim0_size, dim1_size}, dtype);
     }
     return torch::randn({dim0_size, dim1_size}, dtype);
@@ -33,8 +33,8 @@ bool checkPermOf2dTensor(torch::Tensor &a, torch::Tensor &b) {
     return true;
 }
 
-void sortWithinEdgeBuckets(torch::Tensor &rand_tensor, vector<int64_t> &edge_bucket_sizes, int sort_dim) {
-    int64_t offset = 0;
+void sortWithinEdgeBuckets(torch::Tensor &rand_tensor, vector<int32_t> &edge_bucket_sizes, int sort_dim) {
+    int32_t offset = 0;
     for (auto itr = edge_bucket_sizes.begin(); itr != edge_bucket_sizes.end(); itr++) {
         torch::Tensor edge_bucket = rand_tensor.slice(0, offset, offset + *itr);
         edge_bucket.copy_(edge_bucket.index_select(0, torch::argsort(edge_bucket.select(1, sort_dim))));
@@ -50,7 +50,7 @@ bool sortEdgesSrcDest(vector<int> &edge1, vector<int> &edge2) {
     return false;
 }
 
-vector<int64_t> partitionEdges(torch::Tensor &edges, int num_partitions, int num_nodes) {
+vector<int32_t> partitionEdges(torch::Tensor &edges, int num_partitions, int num_nodes) {
     vector<vector<int>> edges_vec(edges.size(0), vector<int>(edges.size(1)));
     for (int i = 0; i < edges_vec.size(); i++) {
         for (int j = 0; j < edges_vec[i].size(); j++) {
@@ -61,7 +61,7 @@ vector<int64_t> partitionEdges(torch::Tensor &edges, int num_partitions, int num
     int partition_size = ceil(((double)num_nodes) / num_partitions);
     std::pair<int, int> prev(edges_vec[0][0] / partition_size, edges_vec[0][2] / partition_size), cur;
     int count = 1;
-    vector<int64_t> partition_sizes_;
+    vector<int32_t> partition_sizes_;
     for (int i = 1; i < edges_vec.size(); i++) {
         cur = std::pair<int, int>(edges_vec[i][0] / partition_size, edges_vec[i][2] / partition_size);
         if (cur == prev) {
